pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-changelog"
version = "0.1.0"
description = "Generate AI-powered summaries of git commit diffs and store them in git notes"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
keywords = ["git", "ai", "summary", "changelog", "notes"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "litellm>=1.0.0",
    "gitpython>=3.1.30",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]

test = [
    "pytest>=7.0",
    "pytest-mock>=3.10",
    "pytest-cov>=4.0",
]
eval = [
    "deepeval>=3.0.0",
    "ollama>=0.4.0",
]

[project.scripts]
ai-changelog = "ai_changelog_msg.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_changelog_msg"]

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py312,py311,py310
skip_missing_interpreters = true
isolated_build = true

[testenv]
description = Run unit tests with pytest
extras = test
setenv =
    PYTHONPATH = {toxinidir}/src
commands =
    pytest {posargs}

[testenv:coverage]
description = Run tests with coverage reporting
extras = test
commands =
    pytest --cov=ai_changelog_msg --cov-report=html --cov-report=term-missing tests/

[testenv:lint]
description = Fix code style
deps =
    ruff
    black
    isort
    pyink
skip_install = true
commands =
    ruff check --fix src/ tests/
    black src/ tests/
    isort src/ tests/
    pyink src/ tests/

[testenv:lint-check]
description = Check code style
deps =
    ruff
    black
    isort
    pyink
skip_install = true
commands =
    ruff check src/ tests/
    black --check src/ tests/
    isort --check src/ tests/
    pyink --check src/ tests/
"""

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "requests.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --tb=short --cov=ai_changelog_msg --cov-report=html --cov-report=term-missing --doctest-modules"
testpaths = ["tests", "src/ai_changelog_msg"]
markers = [
    "unit: Unit tests for individual modules",
    "integration: Integration tests",
    "slow: Tests that take a long time to run",
]
